fix: export embeddings utils
