+++ packages/core/src/embeddings/index.ts
@@ -2,4 +2,10 @@ export { BaseEmbedding, batchEmbeddings } from "./base";
 export type { BaseEmbeddingOptions, EmbeddingInfo } from "./base";
 export { MultiModalEmbedding } from "./muti-model";
 export { truncateMaxTokens } from "./tokenizer";
-export { DEFAULT_SIMILARITY_TOP_K, SimilarityType, similarity } from "./utils";
+export {
+  DEFAULT_SIMILARITY_TOP_K,
+  SimilarityType,
+  getTopKEmbeddings,
+  getTopKMMREmbeddings,
+  similarity,
+} from "./utils";
