+++ packages/llamaindex/src/vector-store/SimpleVectorStore.ts
@@ -1,8 +1,11 @@
 import type { BaseEmbedding } from "@llamaindex/core/embeddings";
+import {
+  getTopKEmbeddings,
+  getTopKMMREmbeddings,
+} from "@llamaindex/core/embeddings";
 import { DEFAULT_PERSIST_DIR } from "@llamaindex/core/global";
 import type { BaseNode } from "@llamaindex/core/schema";
 import { fs, path } from "@llamaindex/env";
-import { getTopKEmbeddings, getTopKMMREmbeddings } from "../internal/utils.js";
 import { exists } from "../storage/FileSystem.js";
 import {
   BaseVectorStore,
